initial wayland support

+++ src/main/scala/msg/Event.scala
@@ -5,11 +5,11 @@ package windows.msg
 
 sealed trait Event
 
-case class ButtonPressEvent(window: Int, button: Int) extends Event
-case class ButtonReleaseEvent(window: Int, button: Int) extends Event
-case class KeyPressEvent(window: Int, key: Int) extends Event
-case class KeyReleaseEvent(window: Int, key: Int) extends Event
-case class MotionNotifyEvent(window: Int) extends Event
-case class MapRequestEvent(window: Int) extends Event
-case class MapNotifyEvent(window: Int) extends Event
-case class UnmapNotifyEvent(window: Int) extends Event
+case class ButtonPressEvent(window: Long, button: Int) extends Event
+case class ButtonReleaseEvent(window: Long, button: Int) extends Event
+case class KeyPressEvent(window: Long, key: Int) extends Event
+case class KeyReleaseEvent(window: Long, key: Int) extends Event
+case class MotionNotifyEvent(window: Long) extends Event
+case class MapRequestEvent(window: Long) extends Event
+case class MapNotifyEvent(window: Long) extends Event
+case class UnmapNotifyEvent(window: Long) extends Event
